Added distance estimation method using target area. Also made it more accurate

+++ research-2019/src/main/java/frc/util/snail_vision/Target.java
@@ -4,10 +4,10 @@ public class Target{
     double[] TARGET_HEIGHT;   
     double[] DESIRED_DISTANCE; // Array so that if a target has multiple levels of height
     
-    double AREA_PERCENT_MEASUREMENTS;
+    int AREA_PERCENT_MEASUREMENTS;
     double[] AREA_TO_DISTANCE = {};
 
-    public Target(double[] targetHeight, double[] desiredDistance, double areaPercentMeasurements, double[] areaToDistance){
+    public Target(double[] targetHeight, double[] desiredDistance, int areaPercentMeasurements, double[] areaToDistance){
         TARGET_HEIGHT = targetHeight;
         DESIRED_DISTANCE = desiredDistance;
         AREA_PERCENT_MEASUREMENTS = areaPercentMeasurements;
